+++ src/conversion/prettify/PrettifyOptions.ts
@@ -1,9 +1,14 @@
+/**
+ * Options for `prettifyPromptbookString` function
+ */
 export type PrettifyOptions = {
+    /***
+     * If true, adds Mermaid graph to the Promptbook string
+     */
     isGraphAdded?: boolean;
 
+    /**
+     * If true, the string is prettifyed as markdown
+     */
     isPrettifyed?: boolean;
 };
-
-/**
- * TODO: !!! Annotate all
- */