+++ src/execution/ExecutionTools.ts
@@ -26,3 +26,7 @@ export type ExecutionTools = {
      */
     userInterface: UserInterfaceTools;
 };
+
+/**
+ * TODO: !!!! Allow more
+ */